+++ src/Digdir.Domain.Dialogporten.Application/Externals/IResourceRegistry.cs
@@ -1,10 +1,12 @@
-﻿namespace Digdir.Domain.Dialogporten.Application.Externals;
+namespace Digdir.Domain.Dialogporten.Application.Externals;
 
 public interface IResourceRegistry
 {
     Task<IReadOnlyCollection<ServiceResourceInformation>> GetResourceInformationForOrg(string orgNumber, CancellationToken cancellationToken);
     Task<ServiceResourceInformation?> GetResourceInformation(string serviceResourceId, CancellationToken cancellationToken);
     IAsyncEnumerable<List<UpdatedSubjectResource>> GetUpdatedSubjectResources(DateTimeOffset since, int batchSize,
+        CancellationToken cancellationToken);
+    Task<IReadOnlyCollection<UpdatedResourcePolicyInformation>> GetUpdatedResourcePolicyInformation(DateTimeOffset since, int numberOfConcurrentRequests,
         CancellationToken cancellationToken);
 }
 
@@ -24,3 +26,4 @@ public sealed record ServiceResourceInformation
 
 
 public sealed record UpdatedSubjectResource(Uri SubjectUrn, Uri ResourceUrn, DateTimeOffset UpdatedAt, bool Deleted);
+public sealed record UpdatedResourcePolicyInformation(Uri ResourceUrn, int MinimumAuthenticationLevel, DateTimeOffset UpdatedAt);