+++ src/Digdir.Domain.Dialogporten.Janitor/Commands.cs
@@ -1,5 +1,6 @@
 using Cocona;
-using Digdir.Domain.Dialogporten.Application.Features.V1.ResourceRegistry.Commands.Synchronize;
+using Digdir.Domain.Dialogporten.Application.Features.V1.ResourceRegistry.Commands.SyncPolicy;
+using Digdir.Domain.Dialogporten.Application.Features.V1.ResourceRegistry.Commands.SyncSubjectMap;
 using MediatR;
 
 namespace Digdir.Domain.Dialogporten.Janitor;
@@ -16,7 +17,22 @@ internal static class Commands
             =>
             {
                 var result = await application.Send(
-                    new SynchronizeSubjectResourceMappingsCommand { Since = since, BatchSize = batchSize },
+                    new SyncSubjectMapCommand { Since = since, BatchSize = batchSize },
+                    ctx.CancellationToken);
+                return result.Match(
+                    success => 0,
+                    validationError => -1);
+            });
+
+        app.AddCommand("sync-resource-policy-information", async (
+                [FromService] CoconaAppContext ctx,
+                [FromService] ISender application,
+                [Option('s')] DateTimeOffset? since,
+                [Option('c')] int? numberOfConcurrentRequests)
+            =>
+            {
+                var result = await application.Send(
+                    new SyncPolicyCommand { Since = since, NumberOfConcurrentRequests = numberOfConcurrentRequests },
                     ctx.CancellationToken);
                 return result.Match(
                     success => 0,