+++ src/Digdir.Domain.Dialogporten.WebApi/OpenApiDocumentExtensions.cs
@@ -7,45 +7,6 @@ namespace Digdir.Domain.Dialogporten.WebApi;
 public static class OpenApiDocumentExtensions
 {
     /// <summary>
-    /// This is a workaround for a bug/weird behaviour in FastEndpoints where examples
-    /// to the RequestExamples array are not added to the OpenAPI document.
-    /// Each SwaggerConfig file implements a GetExample method that returns the example
-    ///
-    /// This sets for those who have a request body.
-    /// </summary>
-    /// <param name="openApiDocument"></param>
-    public static void ReplaceRequestExampleBodies(this OpenApiDocument openApiDocument)
-    {
-        foreach (var path in openApiDocument.Paths)
-        {
-            foreach (var openApiOperation in path.Value.Values)
-            {
-                openApiOperation.ReplaceRequestExampleBody();
-            }
-        }
-    }
-
-    private static void ReplaceRequestExampleBody(this OpenApiOperation openApiOperation)
-    {
-        if (openApiOperation.RequestBody?.Content == null)
-        {
-            return;
-        }
-
-        var operationId = openApiOperation.OperationId;
-
-        // TEMP hard coding of operationId, there is only one endpoint with a request body example
-        // More to follow, make look up function based on operationId
-        if (operationId != "V1ServiceOwnerDialogsUpdate_Dialog") return;
-
-        foreach (var (_, value) in openApiOperation.RequestBody.Content)
-        {
-            var example = UpdateDialogSwaggerConfig.GetExample();
-            value.Example = example;
-        }
-    }
-
-    /// <summary>
     /// When generating ProblemDetails and ProblemDetails_Error, there is a bug/weird behavior in NSwag or FastEndpoints
     /// which results in certain 'Description' properties being generated when running on f.ex. MacOS,
     /// but not when running on the Ubuntu GitHub Actions runner. This leads to the OpenAPI swagger snapshot test
@@ -77,6 +38,30 @@ public static class OpenApiDocumentExtensions
                 {
                     property.Value.Description = null;
                 }
+            }
+        }
+    }
+
+    public static void MakeCollectionsNullable(this OpenApiDocument openApiDocument)
+    {
+        foreach (var schema in openApiDocument.Components.Schemas.Values)
+        {
+            MakeCollectionsNullable(schema);
+        }
+    }
+
+    private static void MakeCollectionsNullable(JsonSchema schema)
+    {
+        if (schema.Properties == null)
+        {
+            return;
+        }
+
+        foreach (var property in schema.Properties.Values)
+        {
+            if (property.Type.HasFlag(JsonObjectType.Array))
+            {
+                property.IsNullableRaw = true;
             }
         }
     }